+++ src/hasCompareFunction.mjs
@@ -1,0 +1,3 @@
+export function hasComparatorFunction(node) {
+  return typeof node.config.modelComparatorFn === 'function';
+}
